Nordic peripheral consistency
Issue #39

Top level Makefile:
bugfix: nordic/tests called multiple times

Event handlers:
  Use size_t for length parameter(s)
  The void* context is always the last parameter.

device drivers:
  spim spim_control_block_t ctor added for bettrer readability.
  spim, spis event type and handler type is now common
  spis implements double buffering
  twim twim_control_block_t ctor added
  twis twis_control_block_t ctor added
  rename rx_length, tx_length -> mosi/miso for SPIS

Add struct spis_buffer_t and spis_buffer_t::buffer_t
for handling the DMA transfer address, length.

spi_test/main.cc
  Uses std::array as buffer implementation rather
    than the native C array types.
  Add double-buffered testing to spi_test.

+++ nordic/peripherals/usart.cc
@@ -110,8 +110,7 @@ struct usart_control_block_t
     usart_control_block_t& operator=(usart_control_block_t const&)  = delete;
     usart_control_block_t& operator=(usart_control_block_t&&)       = delete;
 
-    usart_control_block_t(uint32_t  base_address,
-                          IRQn_Type irq_no)
+    usart_control_block_t(uintptr_t base_address, IRQn_Type irq_no)
     :   usart_registers(reinterpret_cast<NRF_UARTE_Type *>(base_address)),
         irq_type(irq_no),
         handler(nullptr),
